@@ -16,8 +16,8 @@
     "@sveltejs/adapter-auto": "^3.2.1",
     "@sveltejs/adapter-node": "^5.0.1",
     "@sveltejs/adapter-static": "^3.0.1",
-    "@sveltejs/kit": "^2.5.10",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tabler/icons-svelte": "3.5.0",
     "@tailwindcss/typography": "0.5.13",
     "@types/node": "20.14.2",
@@ -28,17 +28,17 @@
     "autoprefixer": "10.4.19",
     "daisyui": "3.9.3",
     "eslint": "^8.56.0",
-    "eslint-plugin-svelte": "2.39.2",
+    "eslint-plugin-svelte": "^2.45.1",
     "jsonwebtoken-esm": "2.0.1",
     "postcss": "8.4.38",
     "qrcode": "1.5.3",
     "socket.io-client": "4.7.5",
-    "svelte": "4.2.18",
-    "svelte-check": "3.8.0",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "svelte-highlight": "7.6.1",
     "tailwindcss": "3.4.4",
     "tslib": "2.6.3",
-    "typescript": "^5.2.2",
-    "vite": "^5.2.13"
+    "typescript": "^5.5.0",
+    "vite": "^5.4.4"
   }
 }
