@@ -1,4 +1,6 @@
 <script lang="ts">
+  import { run } from 'svelte/legacy';
+
   import { gameState } from '$lib/store'
 
   import { manageGame } from '$lib/infrastructure/websockets/admin/adminWSActions'
@@ -12,29 +14,14 @@
   const colors = ['#f6c048', '#b18bec', '#48dbfb', '#1dd1a1']
 
   // Seconds left logic to extract
-  let dateNow = Date.now()
+  let dateNow = $state(Date.now())
   setInterval(() => {
     dateNow = Date.now()
   }, 1000)
 
-  $: questionSeconds = $gameState?.currentQuestion?.question.time
-  $: questionStart = new Date($gameState?.gameStatus?.currentQuestionStart)
 
-  $: secondsLeft =
-    questionSeconds - Math.floor((dateNow - questionStart) / 1000)
 
-  let currentQuestionTimeout = false
-  $: if (secondsLeft <= 0) {
-    currentQuestionTimeout = true
-  }
-  $: if (currentQuestionTimeout) {
-    manageGame(
-      socket,
-      game.gameId,
-      'GAME_QUESTION_RESULTS',
-      $gameState.currentQuestion.question.questionId,
-    )
-  }
+  let currentQuestionTimeout = $state(false)
 
   let unsubscribe = gameState.subscribe((gameStateChange) => {
     if (questionStart !== gameStateChange?.currentQuestion?.questionStart) {
@@ -46,8 +33,31 @@
     unsubscribe()
   })
 
-  export let socket: Socket
-  export let game: Game
+  interface Props {
+    socket: Socket;
+    game: Game;
+  }
+
+  let { socket, game }: Props = $props();
+  let questionSeconds = $derived($gameState?.currentQuestion?.question.time)
+  let questionStart = $derived(new Date($gameState?.gameStatus?.currentQuestionStart))
+  let secondsLeft =
+    $derived(questionSeconds - Math.floor((dateNow - questionStart) / 1000))
+  run(() => {
+    if (secondsLeft <= 0) {
+      currentQuestionTimeout = true
+    }
+  });
+  run(() => {
+    if (currentQuestionTimeout) {
+      manageGame(
+        socket,
+        game.gameId,
+        'GAME_QUESTION_RESULTS',
+        $gameState.currentQuestion.question.questionId,
+      )
+    }
+  });
 </script>
 
 <div class="mx-auto w-5/6 text-center">
