@@ -7,6 +7,6 @@
 	<h1 class="text-center text-4xl font-bold">{$page.error?.message}</h1>
 	<button
 		class="mt-7 rounded bg-blue-500 px-4 py-2 font-bold text-white hover:bg-blue-700"
-		on:click={() => goto('/')}>Go Home</button
+		onclick={() => goto('/')}>Go Home</button
 	>
 </div>
