@@ -1,9 +1,16 @@
 <script>
 	import '@fontsource/chakra-petch'
 	import '../app.css'
+	/**
+	 * @typedef {Object} Props
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { children } = $props();
 </script>
 
-<slot />
+{@render children?.()}
 
 <style>
 	:global(body) {
